+++ custom_components/maytag_appliance/__init__.py
@@ -1,4 +1,4 @@
-"""The Whirlpool Appliance integration."""
+"""The Maytag Appliance integration."""
 
 from __future__ import annotations
 
@@ -7,19 +7,19 @@ from homeassistant.const import CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
-from .api import WhirlpoolApiClient
+from .api import MaytagApiClient
 from .const import PLATFORMS
-from .coordinator import WhirlpoolDataCoordinator
+from .coordinator import MaytagDataCoordinator
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
-    """Set up Whirlpool Appliance from a config entry."""
-    client = WhirlpoolApiClient(
+    """Set up Maytag Appliance from a config entry."""
+    client = MaytagApiClient(
         async_get_clientsession(hass),
         entry.data[CONF_USERNAME],
         entry.data[CONF_PASSWORD],
     )
-    coordinator = WhirlpoolDataCoordinator(hass, entry, client)
+    coordinator = MaytagDataCoordinator(hass, entry, client)
     await coordinator.async_config_entry_first_refresh()
 
     entry.runtime_data = coordinator
@@ -28,5 +28,5 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
 
 async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
-    """Unload a Whirlpool Appliance config entry."""
+    """Unload a Maytag Appliance config entry."""
     return await hass.config_entries.async_unload_platforms(entry, PLATFORMS)